Move XCom tests to tests/models/test_xcom.py
Move XCom tests from `tests/models/test_cleartasks.py` to `tests/models/test_xcom.py`
