+++ tests/models/test_cleartasks.py
@@ -17,17 +17,14 @@
 # under the License.
 
 import datetime
-import os
 import unittest
 
 from airflow import settings
-from airflow.models import DAG, TaskInstance as TI, XCom, clear_task_instances
+from airflow.models import DAG, TaskInstance as TI, clear_task_instances
 from airflow.operators.dummy_operator import DummyOperator
-from airflow.utils import timezone
 from airflow.utils.session import create_session
 from airflow.utils.state import State
 from tests.models import DEFAULT_DATE
-from tests.test_utils.config import conf_vars
 
 
 class TestClearTasks(unittest.TestCase):
@@ -236,155 +233,3 @@ class TestClearTasks(unittest.TestCase):
         self.assertEqual(ti2.try_number, 2)
         # try_number (0) + retries(1)
         self.assertEqual(ti2.max_tries, 1)
-
-    @conf_vars({("core", "enable_xcom_pickling"): "False"})
-    def test_xcom_disable_pickle_type(self):
-        json_obj = {"key": "value"}
-        execution_date = timezone.utcnow()
-        key = "xcom_test1"
-        dag_id = "test_dag1"
-        task_id = "test_task1"
-        XCom.set(key=key,
-                 value=json_obj,
-                 dag_id=dag_id,
-                 task_id=task_id,
-                 execution_date=execution_date)
-
-        ret_value = XCom.get_many(key=key,
-                                  dag_ids=dag_id,
-                                  task_ids=task_id,
-                                  execution_date=execution_date).first().value
-
-        self.assertEqual(ret_value, json_obj)
-
-        session = settings.Session()
-        ret_value = session.query(XCom).filter(XCom.key == key, XCom.dag_id == dag_id,
-                                               XCom.task_id == task_id,
-                                               XCom.execution_date == execution_date
-                                               ).first().value
-
-        self.assertEqual(ret_value, json_obj)
-
-    @conf_vars({("core", "enable_xcom_pickling"): "False"})
-    def test_xcom_get_one_disable_pickle_type(self):
-        json_obj = {"key": "value"}
-        execution_date = timezone.utcnow()
-        key = "xcom_test1"
-        dag_id = "test_dag1"
-        task_id = "test_task1"
-        XCom.set(key=key,
-                 value=json_obj,
-                 dag_id=dag_id,
-                 task_id=task_id,
-                 execution_date=execution_date)
-
-        ret_value = XCom.get_one(key=key,
-                                 dag_id=dag_id,
-                                 task_id=task_id,
-                                 execution_date=execution_date)
-
-        self.assertEqual(ret_value, json_obj)
-
-        session = settings.Session()
-        ret_value = session.query(XCom).filter(XCom.key == key, XCom.dag_id == dag_id,
-                                               XCom.task_id == task_id,
-                                               XCom.execution_date == execution_date
-                                               ).first().value
-
-        self.assertEqual(ret_value, json_obj)
-
-    @conf_vars({("core", "enable_xcom_pickling"): "True"})
-    def test_xcom_enable_pickle_type(self):
-        json_obj = {"key": "value"}
-        execution_date = timezone.utcnow()
-        key = "xcom_test2"
-        dag_id = "test_dag2"
-        task_id = "test_task2"
-        XCom.set(key=key,
-                 value=json_obj,
-                 dag_id=dag_id,
-                 task_id=task_id,
-                 execution_date=execution_date)
-
-        ret_value = XCom.get_many(key=key,
-                                  dag_ids=dag_id,
-                                  task_ids=task_id,
-                                  execution_date=execution_date).first().value
-
-        self.assertEqual(ret_value, json_obj)
-
-        session = settings.Session()
-        ret_value = session.query(XCom).filter(XCom.key == key, XCom.dag_id == dag_id,
-                                               XCom.task_id == task_id,
-                                               XCom.execution_date == execution_date
-                                               ).first().value
-
-        self.assertEqual(ret_value, json_obj)
-
-    @conf_vars({("core", "enable_xcom_pickling"): "True"})
-    def test_xcom_get_one_enable_pickle_type(self):
-        json_obj = {"key": "value"}
-        execution_date = timezone.utcnow()
-        key = "xcom_test3"
-        dag_id = "test_dag"
-        task_id = "test_task3"
-        XCom.set(key=key,
-                 value=json_obj,
-                 dag_id=dag_id,
-                 task_id=task_id,
-                 execution_date=execution_date)
-
-        ret_value = XCom.get_one(key=key,
-                                 dag_id=dag_id,
-                                 task_id=task_id,
-                                 execution_date=execution_date)
-
-        self.assertEqual(ret_value, json_obj)
-
-        session = settings.Session()
-        ret_value = session.query(XCom).filter(XCom.key == key, XCom.dag_id == dag_id,
-                                               XCom.task_id == task_id,
-                                               XCom.execution_date == execution_date
-                                               ).first().value
-
-        self.assertEqual(ret_value, json_obj)
-
-    @conf_vars({("core", "xcom_enable_pickling"): "False"})
-    def test_xcom_disable_pickle_type_fail_on_non_json(self):
-        class PickleRce:
-            def __reduce__(self):
-                return os.system, ("ls -alt",)
-
-        self.assertRaises(TypeError, XCom.set,
-                          key="xcom_test3",
-                          value=PickleRce(),
-                          dag_id="test_dag3",
-                          task_id="test_task3",
-                          execution_date=timezone.utcnow())
-
-    @conf_vars({("core", "xcom_enable_pickling"): "True"})
-    def test_xcom_get_many(self):
-        json_obj = {"key": "value"}
-        execution_date = timezone.utcnow()
-        key = "xcom_test4"
-        dag_id1 = "test_dag4"
-        task_id1 = "test_task4"
-        dag_id2 = "test_dag5"
-        task_id2 = "test_task5"
-
-        XCom.set(key=key,
-                 value=json_obj,
-                 dag_id=dag_id1,
-                 task_id=task_id1,
-                 execution_date=execution_date)
-
-        XCom.set(key=key,
-                 value=json_obj,
-                 dag_id=dag_id2,
-                 task_id=task_id2,
-                 execution_date=execution_date)
-
-        results = XCom.get_many(key=key, execution_date=execution_date)
-
-        for result in results:
-            self.assertEqual(result.value, json_obj)
